Exception specification fixes in Python/Ruby, silence 'no symbols' in SSL/ODBC/MySQL/Postgres modules, better exception messages, int/size_t adjustements

+++ src/C++/FieldMap.cpp
@@ -230,11 +230,11 @@ std::string& FieldMap::calculateString( std::string& result ) const
   return result;
 }
 
-int FieldMap::calculateLength( int beginStringField,
+size_t FieldMap::calculateLength( int beginStringField,
                                int bodyLengthField,
                                int checkSumField ) const
 {
-  int result = 0;
+  size_t result = 0;
   Fields::const_iterator i;
   for ( i = m_fields.begin(); i != m_fields.end(); ++i )
   {
